// Backend/src/model/daftarEvaluatorPermohonan.ts
export interface EvaluatorPermohonanData{
    id_evaluator_permohonan: number;
    id_permohonan: number;
    id_evaluator: number;
    id_operator: number;
    tanggal_disposisi: string;
    tautan_disposisi: string;
    tanggal_pengecekan_evaluator: string;
    catatan_evaluasi: string;
    deleted_at: string;
}